Authorization to Routes

+++ src/services/semestreService.js
@@ -1,20 +1,22 @@
 import api from "./api.js";
+import auth from "./authorization";
 export default {
   get: () => {
-    return api.get("/semestres");
+    return api.get("/semestres", auth);
   },
   store: (obj) => {
-    return api.post("/semestres", obj);
+    return api.post("/semestres", obj, auth);
   },
   destroy: (obj) => {
-    return api.delete(`/semestres/${obj}`);
+    return api.delete(`/semestres/${obj}`, auth);
   },
   getOne: (obj) => {
-    return api.get(`/semestres/${obj}`);
+    return api.get(`/semestres/${obj}`, auth);
   },
   update: (obj) => {
-    return api.patch(`/semestres/${obj.id}`, {
+    var dataToUpdate = {
       end_date: obj.end_date,
-    });
+    };
+    return api.put(`/semestres/${obj.id}`, dataToUpdate, auth);
   },
 };
